% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_config.R
\name{cloud_config}
\alias{cloud_config}
\alias{cc_bootcmd}
\alias{cc_users}
\alias{cc_packages}
\alias{cc_write_files}
\alias{cc_runcmd}
\alias{cc_power_state}
\title{Cloud-Config YAML File}
\usage{
cc_bootcmd(boot_lines = NULL)

cc_users(ssh_path = NULL)

cc_packages(pkgs = NULL)

cc_write_files(write_path = NULL, content = NULL)

cc_runcmd(cmd_lines = NULL)

cc_power_state(reboot = TRUE)
}
\arguments{
\item{boot_lines}{character vector of commands to run at boot}

\item{ssh_path}{path to desired ssh public key (for each user)}

\item{pkgs}{apt-get packages to install}

\item{write_path}{path on the remote server to write new file}

\item{content}{contents of new file to write on remote server}

\item{cmd_lines}{character vector of commands to run after remote server up and running}

\item{reboot}{boolean. TRUE (default) to reboot cloud server after init}
}
\description{
Functions to generate the cloud-config file required to initialize a new cloud-r server.
}
\section{Functions}{
\itemize{
\item \code{cc_bootcmd}: bootcmd

\item \code{cc_users}: user block

\item \code{cc_packages}: apt-get packages

\item \code{cc_write_files}: write files on server

\item \code{cc_runcmd}: run commands

\item \code{cc_power_state}: used for reboot
}}

