% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_github.R
\name{connect_github}
\alias{connect_github}
\title{Connect Server to Github}
\usage{
connect_github(ptoken = NULL)
}
\arguments{
\item{ptoken}{An optional personal access token.
If not provided, will default to system var GITHUB_PAT}
}
\value{
Boolean
}
\description{
Function to connect to github via personal access token, explicitly
provided or through use of environmental variable \code{GITHUB_PAT}.
}
