% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cloud_file.R
\name{cc_r}
\alias{cc_r}
\alias{cc_rstudio_server}
\title{Generate Cloud-Config for Parallel R}
\usage{
cc_r()

cc_rstudio_server()
}
\description{
Functions to generate the cloud-config file that is used to launch single/multi cluster R nodes in the cloud.
}
\section{Functions}{
\itemize{
\item \code{cc_r}: for complete RStudio server environment and opencpu server.

\item \code{cc_rstudio_server}: for complete RStudio server environment and opencpu server.
}}

