% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{helper_funs}
\alias{helper_funs}
\alias{get_pubkey}
\alias{get_template}
\alias{get_os}
\title{Helper Functions}
\usage{
get_pubkey(pubkey_path = NULL, choice = NULL)

get_template()

get_os()
}
\arguments{
\item{pubkey_path}{Path to .ssh file in the system that has entries for the public key}

\item{choice}{TBD}
}
\description{
Read in template cloud config or get public key
}
\section{Functions}{
\itemize{
\item \code{get_pubkey}: Helper function that returns public key stored on local system

\item \code{get_template}: Helper function that returns public key stored on local system

\item \code{get_os}: helper function to lookup OS
}}

