% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCAX.R
\name{loadCAX_xls}
\alias{loadCAX_xls}
\title{Loads the most recent CAX download XLS file in the specified folder}
\usage{
loadCAX_xls(folder = "../data-raw/CAX")
}
\arguments{
\item{folder}{String. Default is /data-raw/CAX}
}
\value{
list
}
\description{
Loads the most recent CAX download XLS file in the specified folder
}
\examples{
CAX.tables <- loadCAX_xls()

}
