% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cloud.R
\name{make_cloud}
\alias{make_cloud}
\alias{make_cloud_init}
\alias{read_pubkey}
\title{Generate Cloud-Init File}
\usage{
read_pubkey(pubkey_path = NULL)

make_cloud_init(pubkey_path = NULL, usr = "ruser")
}
\arguments{
\item{pubkey_path}{Path to .ssh file in the system that has entries for the public key}

\item{usr}{A string of length one naming the user to create on the server at launch. Defaults to 'ruser'}
}
\value{
Prints cloud-init file on console to use with DigitalOceans web UI to launch server
}
\description{
Function to generate a cloud-init file that can be used along with the digitalocean API
to fully launch and depeloy a full scale R environment. This function also references
files within this package that can be customized. These files are run on init, and include
an R script to install basic packages, and another file that can contain arbitrary code to run.
}
\section{Functions}{
\itemize{
\item \code{read_pubkey}: Helper function that returns public key stored on local system

\item \code{make_cloud_init}: Generates cloud_init file on console for now.
}}
\examples{
# Not tested for systems other than windows
# To ensure public key is correct, you can run this to see:
read_pubkey()

#otherwise, the following function call will print the cloud init file on console
make_cloud_init()
}

