% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cloud_init.R
\name{make_cloud_init}
\alias{get_inst_lines}
\alias{get_onboot_vars}
\alias{get_template}
\alias{get_wfchunk}
\alias{make_cloud_init}
\alias{read_pubkey}
\title{Generate Cloud-Init File}
\usage{
make_cloud_init(pubkey_path = NULL, usr = NULL, init_rfiles = NULL, ...)

read_pubkey(pubkey_path = NULL)

get_template()

get_wfchunk(init_rfiles = NULL)

get_inst_lines(init_rfiles = NULL)

get_onboot_vars(...)
}
\arguments{
\item{pubkey_path}{Path to .ssh file in the system that has entries for the public key}

\item{usr}{A string of length one naming the user to create on the server at launch. Defaults to 'ruser'}

\item{init_rfiles}{An optional character vector naming R files that should be run on the server to
setup the environment. See details for important information regarding parameter.}

\item{...}{Optional named arguments representing server env vars to set on boot}
}
\value{
Prints cloud-init file on console to use with DigitalOceans web UI to launch server
}
\description{
Function to generate a cloud-init file that can be used along with the digitalocean API
to fully launch and depeloy a full scale R environment. This function also references
files within this package that can be customized. These files are run on init, and include
an R script to install basic packages, and another file that can contain arbitrary code to run.
}
\details{
If \code{init_rfiles} is being provided, keep the following in mind:
   - Should be one or more filenames only
   - The files should be built with the package and placed in inst/ext/install_scripts
   - Any additional files required by the script when it's running on the server should be here: inst/ext/install_scripts/local

If \code{usr} is not provided, default username is 'ruser', and password is not set (for now)

If \code{GITHUB_PAT} is not provided as part of \code{...}, the default behavior is to use Sys.getenv.
}
\section{Functions}{
\itemize{
\item \code{make_cloud_init}: Generates cloud_init file on console for now.

\item \code{read_pubkey}: Helper function that returns public key stored on local system

\item \code{get_template}: Helper function that returns public key stored on local system

\item \code{get_wfchunk}: Helper function that constructs the write_files block in cloud-init

\item \code{get_inst_lines}: Helper function used by \code{get_wfchunk}

\item \code{get_onboot_vars}: Helper fn to construct the bootcmd chunk with env vars
}}
\examples{
## Not tested for systems other than windows

# To ensure public key is correct, you can run this to see:
read_pubkey()

# otherwise, the following function call will print the cloud init file on console
make_cloud_init()

# Create cloud init file
pat <- "[PLACEHOLDER]"
rfile <- "init_server.R"
make_cloud_init(GITHUB_PAT = pat, init_rfiles = rfile)


## DEVELOPMENT - INTERNAL USE

# To complete remove write files chunk pass NULL to get_wfchunk. To instruct
# the cloud-init script to create a blank file on the server, pass in as an argument
# a name of a file that does not exist in the inst/install_scripts directory of the
# local package

# get_wfchunk(NULL) # makes wf chunk NULL
# get_wfchunk("BLAH") # create empty file on server

}

