% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_pkgs.R
\name{set_default_pkgs}
\alias{get_os}
\alias{set_default_pkgs}
\title{Set default R}
\usage{
set_default_pkgs(pkgs = NULL)

get_os()
}
\arguments{
\item{pkgs}{character vector of package names to start with R}
}
\value{
boolean
}
\description{
Function to insert line to initialize default packages automatically when R starts.
}
\section{Functions}{
\itemize{
\item \code{get_os}: helper function to lookup OS
}}

